websocket関連の処理を１つのファイルにまとめる
