+++ src/api/chatbot/websocket/__init__.py
@@ -1,4 +1,3 @@
 from .manager import ConnectionManager
-from .handlers import WebSocketHandler
 
-__all__ = ["ConnectionManager", "WebSocketHandler"]
+__all__ = ['ConnectionManager']
