+++ src/api/chatbot/websocket/manager.py
@@ -1,40 +1,85 @@
+import asyncio
+import json
+import logging
 from typing import List
+
 from fastapi import WebSocket
-from .handlers import send_websocket_message
+from chatbot.utils.config import create_logger
+from chatbot.utils.sentiment import sentiment_tagging
+
+logger = create_logger(__name__)
 
 class ConnectionManager:
-    def __init__(self):
+    def __init__(self, agent=None, cosmos_repository=None):
         self.active_connections: List[WebSocket] = []
+        self.agent = agent
+        self.cosmos_repository = cosmos_repository
 
     async def connect(self, websocket: WebSocket, subprotocol: str | None = None):
-        """WebSocket接続を確立
-
-        Args:
-            websocket (WebSocket): WebSocket接続
-            subprotocol (str | None): 使用するサブプロトコル
-        """
         await websocket.accept(subprotocol=subprotocol)
         self.active_connections.append(websocket)
 
     def disconnect(self, websocket: WebSocket):
         self.active_connections.remove(websocket)
 
-    async def send_personal_message(self, message: str, websocket: WebSocket):
-        await websocket.send_text(message)
+    @staticmethod
+    def _split_text(text: str, max_length: int = 50) -> List[str]:
+        """テキストを適切な長さに分割する
 
-    async def broadcast(self, message: str):
-        for connection in self.active_connections:
-            await connection.send_text(message)
+        Args:
+            text (str): 分割する元のテキスト
+            max_length (int, optional): 1メッセージの最大文字数. Defaults to 50.
 
-    async def send_message_to_all(self, message: str, type: str):
-        closed_connections = []
-        for websocket in self.active_connections:
-            try:
-                await send_websocket_message(websocket, "", "assistant", "start")
-                await send_websocket_message(websocket, message, type)
-                await send_websocket_message(websocket, "", "assistant", "end")
-            except RuntimeError:
-                closed_connections.append(websocket)
+        Returns:
+            List[str]: 分割されたメッセージのリスト
+        """
+        lines = text.split("\n")
+        result = []
 
-        for closed_websocket in closed_connections:
-            self.disconnect(closed_websocket)
+        for line in lines:
+            if not line:
+                continue
+
+            sentences = line.split("。")
+            for sentence in sentences:
+                if not sentence:
+                    continue
+
+                if len(sentence) > max_length:
+                    parts = sentence.split("、")
+                    result.extend([f"{p}、" for p in parts[:-1]] + [parts[-1]])
+                else:
+                    result.append(sentence)
+
+        return [f"{msg}。" if i < len(result) - 1 else msg for i, msg in enumerate(result)]
+
+    async def send_message(self, websocket: WebSocket, message: str, role: str, type: str = "", emotion: str = "neutral"):
+        """単一のWebSocketメッセージを送信"""
+        if not websocket:
+            logger.error("Can't send message, WebSocket connection is closed.")
+            return
+        elif type == "" and message == "":
+            logger.error("Can't send message, message is empty.")
+            return
+
+        role = "assistant" if role == "message" else role
+        json_data = json.dumps(
+            {"role": role, "text": message, "emotion": emotion, "type": type},
+            ensure_ascii=False,
+        )
+        await websocket.send_text(json_data)
+        await asyncio.sleep(0.01)
+
+    async def process_and_send_messages(self, text: str, websocket: WebSocket, type: str):
+        """メッセージを処理して送信"""
+        messages = self._split_text(text)
+        messages = [msg for msg in messages if msg.strip()]
+
+        await self.send_message(websocket, "", "assistant", "start")
+
+        for message in messages:
+            sentiment = await sentiment_tagging(message)
+            logger.info(f"[Websocket]Assistant: {sentiment} >> {message}")
+            await self.send_message(websocket, message, "assistant", type, sentiment)
+
+        await self.send_message(websocket, "", "assistant", "end")
